# app/Http/Controllers/TaskController.php
<?php

namespace App\Http\Controllers;

use App\Helpers\Configs;
use App\Helpers\Files;
use App\Helpers\OnlineVideos;
use App\Helpers\DataBase;
use App\Helpers\GenerateStaticJsons;
use App\Helpers\Ftp;
use Illuminate\Http\Request;
use OpenApi\Attributes as OA;

class TaskController extends Controller
{
    public function __construct()
    {
        ini_set('memory_limit', '-1');
        set_time_limit(60 * 60);
    }

    #[OA\Get(
        path: '/tasks',
        summary: 'Listar tarefas',
        description: 'Retorna lista de tarefas administrativas disponíveis',
        tags: ['Admin - Tarefas'],
        security: [['bearerAuth' => []]],
        responses: [
            new OA\Response(response: 200, description: 'Lista de tarefas', content: new OA\JsonContent(type: 'array', items: new OA\Items(type: 'object'))),
            new OA\Response(response: 401, description: 'Não autenticado')
        ]
    )]
    public function index(Request $request)
    {
        /*  Configs::refresh();

        $version = Configs::get("version");
        $last_version = Configs::get("last_version");
        $force = ($request->force ?? 0);
        $logs = [];

        if ($force == 1 || $last_version <> $version) {

            //Teve alterações no banco de dados. Gera os dados novamente

            //Ajusta tamanho dos arquivos, caso tenham novos arquivos
            $logs["refresh_files_size"] = Files::refresh_size();

            //Exporta o banco de dados
            $logs["export_database"] = DataBase::export();


            //Atualiza a versão anterior para ficar igual a atual
            $logs["new_version"] = Configs::set("last_version", $version);
        }

        $data = Configs::get();
        return response()->json(["logs" => $logs, "data" => $data]);*/

        return response()->json([]);
    }

    #[OA\Get(
        path: '/tasks/refresh_files_size',
        summary: 'Recalcular tamanho dos arquivos',
        description: 'Recalcula o tamanho de todos os arquivos de mídia armazenados',
        tags: ['Admin - Tarefas'],
        security: [['bearerAuth' => []]],
        responses: [
            new OA\Response(response: 200, description: 'Tamanhos atualizados'),
            new OA\Response(response: 401, description: 'Não autenticado')
        ]
    )]
    public function refresh_files_size($check_version = true)
    {
        if ($check_version) {
            $version = Configs::get("version");
            $last_version = Configs::get("version_files_size");
            if ($last_version == $version) {
                return;
            }
        }
        $ret = Files::refresh_size();
        Configs::set("version_files_size", $version);
        return $ret;
    }

    #[OA\Get(
        path: '/tasks/refresh_files_duration',
        summary: 'Recalcular duração dos áudios',
        description: 'Recalcula a duração de todos os arquivos de áudio',
        tags: ['Admin - Tarefas'],
        security: [['bearerAuth' => []]],
        responses: [
            new OA\Response(response: 200, description: 'Durações atualizadas'),
            new OA\Response(response: 401, description: 'Não autenticado')
        ]
    )]
    public function refresh_files_duration($check_version = true)
    {
        if ($check_version) {
            $version = Configs::get("version");
            $last_version = Configs::get("version_files_duration");
            if ($last_version == $version) {
                return;
            }
        }
        $ret = Files::refresh_duration();
        Configs::set("version_files_duration", $version);
        return $ret;
    }

    #[OA\Get(
        path: '/tasks/refresh_online_videos',
        summary: 'Recarregar vídeos online',
        description: 'Atualiza dados de vídeos online do YouTube',
        tags: ['Admin - Tarefas'],
        security: [['bearerAuth' => []]],
        responses: [
            new OA\Response(response: 200, description: 'Vídeos atualizados'),
            new OA\Response(response: 401, description: 'Não autenticado')
        ]
    )]
    public function refresh_online_videos()
    {
        $ret = OnlineVideos::refresh();
        if ($ret["status"] == "") {
            $ret = [];
        }
        return $ret;
    }

    #[OA\Get(
        path: '/tasks/refresh_configs',
        summary: 'Recarregar configurações',
        description: 'Recarrega o cache de configurações do sistema',
        tags: ['Admin - Tarefas'],
        security: [['bearerAuth' => []]],
        responses: [
            new OA\Response(response: 200, description: 'Configurações recarregadas'),
            new OA\Response(response: 401, description: 'Não autenticado')
        ]
    )]
    public function refresh_configs()
    {
        $ret = Configs::refresh();
        if ($ret["status"] <> "") {
            $data = Configs::get();
            $ret["data"] = $data;
        } else {
            $ret = [];
        }

        return $ret;
    }

    #[OA\Get(
        path: '/tasks/export_database',
        summary: 'Exportar banco de dados (SQL)',
        description: 'Gera exportação SQL do banco de dados para o desktop app',
        tags: ['Admin - Tarefas'],
        security: [['bearerAuth' => []]],
        responses: [
            new OA\Response(response: 200, description: 'Exportação concluída', content: new OA\JsonContent(type: 'object')),
            new OA\Response(response: 401, description: 'Não autenticado')
        ]
    )]
    public function export_database($check_version = true)
    {
        if ($check_version) {
            $version = Configs::get("version");
            $last_version = Configs::get("version_export_database");
            if ($last_version == $version) {
                return;
            }
        }

        $ret = DataBase::export();
        if ($ret["error"] && $ret["error"] <> "") {
            Configs::set("version_export_database", -1);
        } else {
            Configs::set("version_export_database", $version);
        }
        return $ret;
    }

    #[OA\Get(
        path: '/tasks/export_database_json',
        summary: 'Exportar banco de dados (JSON)',
        description: 'Gera exportação JSON do banco de dados',
        tags: ['Admin - Tarefas'],
        security: [['bearerAuth' => []]],
        responses: [
            new OA\Response(response: 200, description: 'Exportação concluída', content: new OA\JsonContent(type: 'object')),
            new OA\Response(response: 401, description: 'Não autenticado')
        ]
    )]
    public function export_database_json($check_version = true)
    {
        if (request("force") && request("force") == "true") {
            $check_version = false;
        }

        if ($check_version) {
            $version = Configs::get("version");
            $last_version = Configs::get("version_export_database_json");
            if ($last_version == $version) {
                return [];
            }
        }

        $ret = DataBase::export_json();
        if ($check_version) {
            Configs::set("version_export_database_json", $version);
        }
        return $ret;
    }

    #[OA\Get(
        path: '/tasks/send_database_ftp',
        summary: 'Enviar banco via FTP',
        description: 'Envia exportação do banco de dados para os servidores FTP configurados',
        tags: ['Admin - Tarefas'],
        security: [['bearerAuth' => []]],
        responses: [
            new OA\Response(response: 200, description: 'Envio concluído'),
            new OA\Response(response: 401, description: 'Não autenticado')
        ]
    )]
    public function send_database_ftp($check_version = true)
    {
        if ($check_version) {
            $version = Configs::get("version");
            $last_version = Configs::get("version_send_database_ftp");
            if ($last_version == $version) {
                return;
            }
        }

        Files::permissions(config("files.dir"), 0644, 0755);
        $ret = Ftp::send_database();
        Files::permissions(config("files.dir"), 0444, 0555);
        if ($ret["status"] == true) {
            Configs::set("version_send_database_ftp", $version);
        }
        return $ret;
    }

    #[OA\Get(
        path: '/tasks/import_slides',
        summary: 'Importar slides',
        description: 'Importa slides de apresentações de um diretório configurado',
        tags: ['Admin - Tarefas'],
        security: [['bearerAuth' => []]],
        responses: [
            new OA\Response(response: 200, description: 'Importação concluída'),
            new OA\Response(response: 401, description: 'Não autenticado')
        ]
    )]
    public function import_slides()
    {
        $dir = app()->basePath('public') . DIRECTORY_SEPARATOR . 'import' . DIRECTORY_SEPARATOR;

        $files = Files::list_files($dir);

        if (isset($files["error"])) {
            return response()->json($files);
        }

        $log = [];
        foreach ($files as $file) {
            $ret = DataBase::import_file($file["path"]);
            $log[] = ['file' => $file['name'], 'status' => $ret];
        }

        return response()->json($log);
    }

    #[OA\Get(
        path: '/tasks/generate_static_jsons',
        summary: 'Gerar JSONs estáticos do banco',
        description: 'Gera arquivos JSON estáticos (categorias, albums, musics, hinario, collections online) a partir do banco de dados para uso offline pelo app desktop/web. Os arquivos são salvos em public/db/json/ com hash MD5 para versionamento via ETag.',
        tags: ['Admin - Tarefas'],
        security: [['bearerAuth' => []]],
        parameters: [
            new OA\Parameter(
                name: 'force',
                in: 'query',
                schema: new OA\Schema(type: 'string', enum: ['true', 'false']),
                description: 'Força regeneração ignorando version check'
            ),
        ],
        responses: [
            new OA\Response(response: 200, description: 'JSONs gerados com sucesso', content: new OA\JsonContent(
                type: 'object',
                properties: [
                    new OA\Property(property: 'status', type: 'string', example: 'success'),
                    new OA\Property(property: 'files_generated', type: 'integer', example: 42),
                    new OA\Property(property: 'logs', type: 'array', items: new OA\Items(type: 'object')),
                ]
            )),
            new OA\Response(response: 401, description: 'Não autenticado')
        ]
    )]
    public function generate_static_jsons()
    {
        $force = request('force') === 'true';
        $checkVersion = !$force;

        if ($checkVersion) {
            $version = Configs::get("version");
            $lastVersion = Configs::get("version_generate_static_jsons");
            if ($lastVersion == $version) {
                return [];
            }
        }

        $logs = GenerateStaticJsons::generate();
        Configs::set("version_generate_static_jsons", Configs::get("version"));

        return response()->json([
            'status' => 'success',
            'message' => 'JSONs estáticos gerados com sucesso',
            'files_generated' => count($logs),
            'logs' => $logs,
        ]);
    }
}