# storage/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LouvorJA API",
        "description": "API do LouvorJA - Gerenciador de Hinários e Músicas para a Igreja Adventista do Sétimo Dia. Fornece endpoints públicos para consulta de músicas, letras, arquivos e álbuns, além de endpoints administrativos protegidos por autenticação JWT.",
        "contact": {
            "name": "LouvorJA",
            "email": "contato@louvorja.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.louvorja.com.br",
            "description": "Produção"
        },
        {
            "url": "http://localhost:8000",
            "description": "Desenvolvimento"
        }
    ],
    "paths": {
...
        "/tasks": {
            "get": {
                "tags": [
                    "Admin - Tarefas"
                ],
                "summary": "Listar tarefas",
                "description": "Retorna lista de tarefas administrativas disponíveis",
                "operationId": "6a93d54730cde509d735e538ab6800f6",
                "responses": {
                    "200": {
                        "description": "Lista de tarefas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/refresh_files_size": {
            "get": {
                "tags": [
                    "Admin - Tarefas"
                ],
                "summary": "Recalcular tamanho dos arquivos",
                "description": "Recalcula o tamanho de todos os arquivos de mídia armazenados",
                "operationId": "76cfb61bf23c46322fc5d2e15b6e34ee",
                "responses": {
                    "200": {
                        "description": "Tamanhos atualizados"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/refresh_files_duration": {
            "get": {
                "tags": [
                    "Admin - Tarefas"
                ],
                "summary": "Recalcular duração dos áudios",
                "description": "Recalcula a duração de todos os arquivos de áudio",
                "operationId": "3f66d2368ca15bd0814e496e5467f428",
                "responses": {
                    "200": {
                        "description": "Durações atualizadas"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/refresh_online_videos": {
            "get": {
                "tags": [
                    "Admin - Tarefas"
                ],
                "summary": "Recarregar vídeos online",
                "description": "Atualiza dados de vídeos online do YouTube",
                "operationId": "a55f655fab7ec6b584bcbdb3971f0e6c",
                "responses": {
                    "200": {
                        "description": "Vídeos atualizados"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/refresh_configs": {
            "get": {
                "tags": [
                    "Admin - Tarefas"
                ],
                "summary": "Recarregar configurações",
                "description": "Recarrega o cache de configurações do sistema",
                "operationId": "d0fe5cd403bb39812547fbf056bdfa77",
                "responses": {
                    "200": {
                        "description": "Configurações recarregadas"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/export_database": {
            "get": {
                "tags": [
                    "Admin - Tarefas"
                ],
                "summary": "Exportar banco de dados (SQL)",
                "description": "Gera exportação SQL do banco de dados para o desktop app",
                "operationId": "4e7b24b851ca3008fb33be0ec34cc38d",
                "responses": {
                    "200": {
                        "description": "Exportação concluída",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/export_database_json": {
            "get": {
                "tags": [
                    "Admin - Tarefas"
                ],
                "summary": "Exportar banco de dados (JSON)",
                "description": "Gera exportação JSON do banco de dados",
                "operationId": "cc1a7a173f261fe256e695b726a97700",
                "responses": {
                    "200": {
                        "description": "Exportação concluída",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/send_database_ftp": {
            "get": {
                "tags": [
                    "Admin - Tarefas"
                ],
                "summary": "Enviar banco via FTP",
                "description": "Envia exportação do banco de dados para os servidores FTP configurados",
                "operationId": "3dc1b42695f193e478ee3d60081be062",
                "responses": {
                    "200": {
                        "description": "Envio concluído"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/import_slides": {
            "get": {
                "tags": [
                    "Admin - Tarefas"
                ],
                "summary": "Importar slides",
                "description": "Importa slides de apresentações de um diretório configurado",
                "operationId": "56deaad868763b40820db828cc432a9e",
                "responses": {
                    "200": {
                        "description": "Importação concluída"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tasks/generate_static_jsons": {
            "get": {
                "tags": [
                    "Admin - Tarefas"
                ],
                "summary": "Gerar JSONs estáticos do banco",
                "description": "Gera arquivos JSON estáticos (categorias, albums, musics, hinario, collections online) a partir do banco de dados para uso offline pelo app desktop/web. Os arquivos são salvos em public/db/json/ com hash MD5 para versionamento via ETag.",
                "operationId": "b7879db5ff0e29a5598ba721f047519f",
                "parameters": [
                    {
                        "name": "force",
                        "in": "query",
                        "description": "Força regeneração ignorando version check",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true",
                                "false"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSONs gerados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "files_generated": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "logs": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação JWT via header Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}